+++ src/main/java/com/sprint/sb06deokhugamteam01/domain/Comment.java
@@ -2,7 +2,6 @@ package com.sprint.sb06deokhugamteam01.domain;
 
 import jakarta.persistence.*;
 
-import com.sprint.sb06deokhugamteam01.domain.review.Review;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EntityListeners;
