+++ src/main/java/com/sprint/sb06deokhugamteam01/domain/Notification.java
@@ -1,6 +1,5 @@
 package com.sprint.sb06deokhugamteam01.domain;
 
-import com.sprint.sb06deokhugamteam01.domain.review.Review;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EntityListeners;
