+++ src/main/java/com/sprint/sb06deokhugamteam01/dto/review/ReviewSearchCondition.java
@@ -1,4 +1,4 @@
-package com.sprint.sb06deokhugamteam01.domain.review;
+package com.sprint.sb06deokhugamteam01.dto.review;
 
 import lombok.Builder;
 import org.springframework.data.domain.Pageable;
