+++ src/main/java/com/sprint/sb06deokhugamteam01/dto/review/request/ReviewCreateRequest.java
@@ -1,4 +1,4 @@
-package com.sprint.sb06deokhugamteam01.dto.review;
+package com.sprint.sb06deokhugamteam01.dto.review.request;
 
 import jakarta.validation.constraints.*;
 import lombok.Builder;
