+++ src/main/java/com/sprint/sb06deokhugamteam01/repository/review/ReviewLikeRepository.java
@@ -2,14 +2,13 @@ package com.sprint.sb06deokhugamteam01.repository.review;
 
 import com.sprint.sb06deokhugamteam01.domain.ReviewLike;
 import com.sprint.sb06deokhugamteam01.domain.User;
-import com.sprint.sb06deokhugamteam01.domain.review.Review;
+import com.sprint.sb06deokhugamteam01.domain.Review;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 import java.util.Optional;
-import java.util.Set;
 import java.util.UUID;
 
 public interface ReviewLikeRepository extends JpaRepository<ReviewLike, UUID> {
