+++ src/main/java/com/sprint/sb06deokhugamteam01/repository/review/ReviewRepository.java
@@ -2,9 +2,8 @@ package com.sprint.sb06deokhugamteam01.repository.review;
 
 import com.sprint.sb06deokhugamteam01.domain.Book;
 import com.sprint.sb06deokhugamteam01.domain.User;
-import com.sprint.sb06deokhugamteam01.domain.review.Review;
+import com.sprint.sb06deokhugamteam01.domain.Review;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 
 import java.util.List;
 import java.util.UUID;
