+++ src/main/java/com/sprint/sb06deokhugamteam01/repository/review/ReviewRepositoryCustom.java
@@ -1,11 +1,10 @@
 package com.sprint.sb06deokhugamteam01.repository.review;
 
-import com.sprint.sb06deokhugamteam01.domain.review.PopularReviewSearchCondition;
-import com.sprint.sb06deokhugamteam01.domain.review.Review;
-import com.sprint.sb06deokhugamteam01.domain.review.ReviewSearchCondition;
+import com.sprint.sb06deokhugamteam01.dto.review.PopularReviewSearchCondition;
+import com.sprint.sb06deokhugamteam01.domain.Review;
+import com.sprint.sb06deokhugamteam01.dto.review.ReviewSearchCondition;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
-import org.springframework.stereotype.Repository;
 
 public interface ReviewRepositoryCustom {
